move routes/ and scenes/ to components/, move existing components/ to components/ui
