+++ app/src/components/scenes/settings/Settings.js
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import {
   StyleSheet, Text, View, StatusBar,
 } from "react-native";
-import Button from "components/Button";
+import Button from "components/ui/Button";
 import { colors } from "theme";
 
 const styles = StyleSheet.create({
